+++ src/main/java/com/udanchoo/intranet/repository/incentive/EmployeeTargetMappingRepository.java
@@ -7,7 +7,7 @@ import com.udanchoo.intranet.entity.EmployeeTargetMappingEntity;
 
 public interface EmployeeTargetMappingRepository extends JpaRepository<EmployeeTargetMappingEntity, Long>,JpaSpecificationExecutor {
 
-    boolean existsByUserIdAndFinancialYearAndTargetAmount(int userId, String financialYear, double targetAmount);
+    boolean existsByUserIdAndFinancialYearAndTargetAmount(int userId, String financialYear, int targetAmount);
 
     // Add custom query methods if needed
 }
